feat!: rename package and repository to ha-mcp
BREAKING CHANGE: Package name changed from homeassistant-mcp to ha-mcp

This is a complete rename to align the package name with the repository:
- Repository: homeassistant-ai/ha-mcp
- Package: ha-mcp (was homeassistant-mcp)
- Module: ha_mcp (was homeassistant_mcp)
- CLI: ha-mcp (was homeassistant-mcp)

## Migration Guide

### For new installations:
```bash
git clone https://github.com/homeassistant-ai/ha-mcp
cd ha-mcp
uv sync
uv run ha-mcp
```

### For existing users:

1. Update git remote:
```bash
git remote set-url origin https://github.com/homeassistant-ai/ha-mcp
```

2. Update imports in your code:
```python
# Old
from homeassistant_mcp import HomeAssistantClient

# New
from ha_mcp import HomeAssistantClient
```

3. Update CLI command:
```bash
# Old
uv run homeassistant-mcp

# New
uv run ha-mcp
```

## Changes
- Renamed `src/homeassistant_mcp/` → `src/ha_mcp/`
- Updated package name in pyproject.toml
- Updated all Python imports across codebase
- Updated CLI command name
- Updated documentation (README, AGENTS.md)
- Updated repository URLs
- Updated CHANGELOG.md historical references
