+++ .env.example
@@ -24,5 +24,5 @@ DEBUG=false
 LOG_LEVEL=INFO
 
 # Optional: MCP Server Configuration
-MCP_SERVER_NAME=homeassistant-mcp
+MCP_SERVER_NAME=ha-mcp
 MCP_SERVER_VERSION=0.1.0
