+++ fastmcp.json
@@ -1,7 +1,7 @@
 {
   "$schema": "https://gofastmcp.com/public/schemas/fastmcp.json/v1.json",
   "source": {
-    "path": "src/homeassistant_mcp/__main__.py"
+    "path": "src/ha_mcp/__main__.py"
   },
   "deployment": {
     "transport": "stdio",
