+++ src/ha_mcp/__init__.py
