+++ src/ha_mcp/client/__init__.py
