+++ src/ha_mcp/client/rest_client.py
