+++ src/ha_mcp/prompts/enhanced.py
