+++ src/ha_mcp/prompts/manager.py
