+++ src/ha_mcp/resources/manager.py
