+++ src/ha_mcp/server.py
