+++ src/ha_mcp/tools/convenience.py
