+++ src/ha_mcp/tools/device_control.py
