+++ src/ha_mcp/tools/enhanced.py
