+++ src/ha_mcp/tools/registry.py
