+++ src/ha_mcp/tools/smart_search.py
