+++ src/ha_mcp/utils/domain_handlers.py
