+++ src/ha_mcp/utils/operation_manager.py
